Code refactor: create DataElement folder and move some classes there
