+++ source/AMDSDataStream.cpp
@@ -1,7 +1,7 @@
 #include "source/AMDSDataStream.h"
 
 #include "source/ClientRequest/AMDSClientRequestSupport.h"
-#include "source/AMDSEventDataSupport.h"
+#include "source/DataElement/AMDSEventDataSupport.h"
 #include "source/DataHolder/AMDSDataHolderSupport.h"
 #include "source/util/AMDSErrorMonitor.h"
