--- source/AMDSDataTypeDefinitions.h
@@ -1,20 +1,0 @@
-#ifndef AMDSDATATYPEDEFINITIONS_H
-#define AMDSDATATYPEDEFINITIONS_H
-
-namespace AMDSDataTypeDefinitions {
-	enum DataType{
-		Signed8 = 0,
-		Unsigned8 = 1,
-		Signed16 = 2,
-		Unsigned16 = 3,
-		Signed32 = 4,
-		Unsigned32 = 5,
-		Signed64 = 6,
-		Unsigned64 = 7,
-		Float = 8,
-		Double = 9,
-		InvalidType = 10
-	};
-}
-
-#endif // AMDSDATATYPEDEFINITIONS_H
