--- source/AMDSFlatArray.cpp
@@ -1,11 +1,0 @@
-#include "source/AMDSFlatArray.h"
-
-AMDSFlatArray::AMDSFlatArray(AMDSDataTypeDefinitions::DataType dataType, quint32 size)
-{
-	dataType_ = dataType;
-	resizeType(dataType_, size);
-}
-
-AMDSFlatArray::~AMDSFlatArray()
-{
-}
