+++ source/AMDSMain.cpp
@@ -6,8 +6,8 @@
 #include <QSettings>
 
 #include "source/AMDSCentralServer.h"
-#include "source/AMDSEventDataSupport.h"
 #include "source/ClientRequest/AMDSClientRequestSupport.h"
+#include "source/DataElement/AMDSEventDataSupport.h"
 #include "source/DataHolder/AMDSDataHolderSupport.h"
 
 /*
